requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
backports.zoneinfo==0.2.1
click==7.1.2
convertdate==2.3.1
coverage==5.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-autocomplete-light==3.8.2
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-heroku==0.3.1
exceptiongroup==1.2.2
filelock==3.0.12
glob2==0.7
gunicorn==20.0.4
hijri-converter==2.1.1
holidays==0.10.5.2
iniconfig==2.0.0
isort==4.3.21
junit-xml==1.8
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mock==4.0.3
numpy==1.24.4
packaging==24.1
pandas==2.0.3
parso==0.7.0
pbr==5.5.1
pipdeptree==2.3.3
plotly==5.24.1
pluggy==1.5.0
psycopg2==2.8.5
pylint==2.5.3
PyMeeus==0.5.3
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2020.1
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
stevedore==3.3.0
tenacity==9.0.0
time-machine==2.1.0
toml==0.10.1
tomli==2.0.2
tzdata==2024.2
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
whitenoise==5.2.0
wrapt==1.12.1
